arm7 dump, overlay 204 dump

+++ overlays/overlay_93/src/ov93_021B8724.c
@@ -4,7 +4,7 @@
 
 void ov93_021B8724(void* context, u32 param) {
     u32* ptr = *(u32**)context;
-    u16 converted = ov93_021D5444(param);
+    u16 converted = (u16)ov93_021D5444(param);
     u32 field = ptr[0x68 / 4];
     FUN_0200CED4(field, converted);
 }
